test: 클래스별 단위로 테스트 코드 분리
